Error page has been added

+++ WeatherApp/Models/WeatherResponse.cs
@@ -59,16 +59,16 @@ namespace WeatherApp.Models
     public class Main
     {
         public double temp { get; set; }
-        public int pressure { get; set; }
-        public int humidity { get; set; }
+        public double pressure { get; set; }
+        public double humidity { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
     }
 
     public class Wind
     {
-        public int speed { get; set; }
-        public int deg { get; set; }
+        public double speed { get; set; }
+        public double deg { get; set; }
     }
 
     public class Clouds
@@ -110,10 +110,10 @@ namespace WeatherApp.Models
         public int dt { get; set; }
         public Temp temp { get; set; }
         public double pressure { get; set; }
-        public int humidity { get; set; }
+        public double humidity { get; set; }
         public List<Weather> weather { get; set; }
         public double speed { get; set; }
-        public int deg { get; set; }
-        public int clouds { get; set; }
+        public double deg { get; set; }
+        public double clouds { get; set; }
     }
 }
